Conflicts:
	src/cryptohelper/com/GUIController.java

+++ src/cryptohelper/service/DBController.java
@@ -4,9 +4,9 @@ package cryptohelper.service;
 
 import cryptohelper.data.Messaggio;
 import cryptohelper.data.QueryResult;
-import cryptohelper.data.Studente;
 import cryptohelper.data.UserInfo;
 import java.sql.Connection;
+import java.sql.DatabaseMetaData;
 import java.sql.DriverManager;
 import java.sql.ResultSet;
 import java.sql.ResultSetMetaData;
@@ -62,19 +62,19 @@ public class DBController {
         } catch (SQLException e) {
             log.fatal(e.getMessage());
         }
-        log.info("Connesso a:" + dBurl);
+        System.out.println("Connesso a:" + dBurl);
     }
 
     //Chiude la connessione al db
     private void disconnect() throws SQLException {
         st.close();
         conn.close();
-        log.trace("Disconnesso!");
+        System.out.println("Disconnesso!");
     }
 
     //Crea le tabelle del database
     public void createTables() throws SQLException {
-        connect();
+        //connect();
         //Query tabella Studenti
         String queryStudenti = "CREATE TABLE Studenti"
                 + "("
@@ -90,6 +90,10 @@ public class DBController {
                 + "("
                 + "ID INTEGER not null primary key "
                 + "GENERATED ALWAYS AS IDENTITY (START WITH 1, INCREMENT BY 1),"
+                + "ID_MITTENTE INTEGER,"
+                + "ID_DESTINATARIO INTEGER,"
+                + "FOREIGN KEY(ID_MITTENTE) REFERENCES STUDENTI(ID),"
+                + "FOREIGN KEY(ID_DESTINATARIO) REFERENCES STUDENTI(ID),"
                 + "Testo VARCHAR(2048),"
                 + "TestoCifrato VARCHAR(2048),"
                 + "Lingua VARCHAR(32),"
@@ -117,48 +121,42 @@ public class DBController {
                 + "FOREIGN KEY(ID_PARTNER) REFERENCES STUDENTI(ID),"
                 + "FOREIGN KEY(ID_SDC) REFERENCES SistemiCifratura(ID)"
                 + ")";
-        String queryMessaggiInviati = "CREATE TABLE MessaggiInviati"
-                + "("
-                + "ID_MITTENTE INTEGER not null,"
-                + "ID_DESTINATARIO INTEGER not null,"
-                + "ID_MESSAGGIO INTEGER not null,"
-                + "PRIMARY KEY(ID_MITTENTE, ID_DESTINATARIO, ID_MESSAGGIO),"
-                + "FOREIGN KEY(ID_MITTENTE) REFERENCES STUDENTI(ID),"
-                + "FOREIGN KEY(ID_DESTINATARIO) REFERENCES STUDENTI(ID),"
-                + "FOREIGN KEY(ID_MESSAGGIO) REFERENCES Messaggi(ID)"
-                + ")";
+        connect();
+        //Non bisogna piu comentare le tabele per far funzionare il Test
         try {
-
-            log.info("DA TOGLIERE I COMENTI DI ST.EXECUTE(QUERRY) QUANDO SI ESEGUE TESTDBCONTROLLER PER LA SECONDA VOLTA!!!!!!!");
             //drop di tutte le tabelle esistenti
-            st.execute("DROP TABLE MessaggiInviati");
-            log.info("Tabella MessaggiInviati eliminata!");
-            st.execute("DROP TABLE SDCPartners");
-            log.info("Tabella SDCPartners eliminata!");
-            st.execute("DROP TABLE SistemiCifratura");
-            log.info("Tabella SistemiCifratura eliminata!");
-            st.execute("DROP TABLE Messaggi");
-            log.info("Tabella Messaggi eliminata!");
-            st.execute("DROP TABLE Studenti");
-            log.info("Tabella Studenti eliminata!");
-
+            if (isTableExist("SDCPartners")) {
+                st.execute("DROP TABLE SDCPartners");
+                System.out.println("Tabella SDCPartners eliminata!");
+            }
+            if (isTableExist("SistemiCifratura")) {
+                st.execute("DROP TABLE SistemiCifratura");
+                System.out.println("Tabella SistemiCifratura eliminata!");
+            }
+            if (isTableExist("Studenti")) {
+                st.execute("DROP TABLE Studenti");
+                System.out.println("Tabella Studenti eliminata!");
+            }
+            if (isTableExist("Messaggi")) {
+                st.execute("DROP TABLE Messaggi");
+                System.out.println("Tabella Messaggi eliminata!");
+            }
+            
             //creazione tabelle
             st.executeUpdate(queryStudenti);
-            log.info("Tabella Studenti creata!");
+            System.out.println("Tabella Studenti creata!");
             st.executeUpdate(queryMessaggi);
-            log.info("Tabella Messaggi creata!");
+            System.out.println("Tabella Messaggi creata!");
             st.executeUpdate(querySistemiCifratura);
-            log.info("Tabella SistemiCifratura creata!");
+            System.out.println("Tabella SistemiCifratura creata!");
             st.executeUpdate(querySDCPartners);
-            log.info("Tabella SDCPartners creata!");
-            st.executeUpdate(queryMessaggiInviati);
-            log.info("Tabella MessaggiInviati creata!");
-
+            System.out.println("Tabella SDCPartners creata!");
         } catch (SQLException e) {
             log.fatal(e.getMessage());
         } finally {
             disconnect();
         }
+
     }
 
     //Per inserimenti
@@ -166,7 +164,7 @@ public class DBController {
         connect();
         try {
             st.executeUpdate(query);
-            log.info("Query eseguita correttamente!");
+            System.out.println("Query eseguita correttamente!");
         } catch (SQLException e) {
             log.fatal(e.getMessage());
         } finally {
@@ -198,20 +196,21 @@ public class DBController {
         }
         return new QueryResult(resultList);
     }
+    /*
+     public ResultSet executeQuery1(String querry) throws SQLException {
 
-    public ResultSet executeQuery1(String querry) throws SQLException {
-
-        ResultSet resultSet = null;
-        connect();
-        try {
-            resultSet = st.executeQuery(querry);
-        } catch (SQLException e) {
-            System.out.println(e.getMessage());
-        } finally {
-            disconnect();
-        }
-        return resultSet;
-    }
+     ResultSet resultSet = null;
+     connect();
+     try {
+     resultSet = st.executeQuery(querry);
+     } catch (SQLException e) {
+     System.out.println(e.getMessage());
+     } finally {
+     disconnect();
+     }
+     return resultSet;
+     }
+     */
 
     public int executeUpdateAndReturnKey(String querry) throws SQLException {
         connect();
@@ -219,7 +218,7 @@ public class DBController {
         try {
             st.executeUpdate(querry, Statement.RETURN_GENERATED_KEYS);
             rs = st.getGeneratedKeys();
-            log.info("Query: " + querry + " - eseguita correttamente!");
+            System.out.println("Query: " + querry + " - eseguita correttamente!");
             while (rs.next()) {
                 result = rs.getInt(1);
             }
@@ -231,6 +230,7 @@ public class DBController {
         return result;
     }
 
+    //TO-DO
     //Preleva il messaggio indicato come parametro dal db
     public Messaggio getMessaggio(int id) throws SQLException {
         connect();
@@ -250,27 +250,26 @@ public class DBController {
         disconnect();
         return result;
     }
+    /*
+     public ArrayList<UserInfo> getDestinatari() throws SQLException {
+     connect();
 
-    
-    public ArrayList<UserInfo> getDestinatari() throws SQLException {
-        connect();
-
-        ArrayList<UserInfo> ui = new ArrayList<>();
-        String querry = "SELECT * FROM STUDENTI";
-        try {
-            rs = st.executeQuery(querry);
-            while (rs.next()) {
-                UserInfo tempU = new UserInfo(rs.getInt("id"), rs.getString("nome"), rs.getString("cognome"));
-                ui.add(tempU);
-            }
-        } catch (SQLException e) {
-            System.out.println(e.getMessage());
-        } finally {
-            disconnect();
-        }
-        System.out.println("Extracted:" + ui.toString());
-        return ui;
-    }
+     ArrayList<UserInfo> ui = new ArrayList<>();
+     String querry = "SELECT * FROM STUDENTI";
+     try {
+     rs = st.executeQuery(querry);
+     while (rs.next()) {
+     UserInfo tempU = new UserInfo(rs.getInt("id"), rs.getString("nome"), rs.getString("cognome"));
+     ui.add(tempU);
+     }
+     } catch (SQLException e) {
+     System.out.println(e.getMessage());
+     } finally {
+     disconnect();
+     }
+     System.out.println("Extracted:" + ui.toString());
+     return ui;
+     }*/
 
     public UserInfo getUserInfo(int id) throws SQLException {
         connect();
@@ -329,5 +328,17 @@ public class DBController {
             System.out.println("Record NOT EDITED with success");
         }
         return (edit == 1 ? true : false);
+    }
+
+    private boolean isTableExist(String sTablename) throws SQLException {
+        DatabaseMetaData dbmd = conn.getMetaData();
+        ResultSet rs = dbmd.getTables(null, null, sTablename.toUpperCase(), null);
+        if (rs.next()) {
+            System.out.println("Table " + rs.getString("TABLE_NAME") + "already exists !!");
+            return true;
+        } else {
+            System.out.println("creating table: " + sTablename);
+            return false;
+        }
     }
 }
